Add coverage report to GitHub workflow
This commit introduces a new step to the GitHub actions workflow that generates a code coverage report. The change also includes adjustments to the Makefile to include coverage commands, and the .envrc file has been tweaked to interact with .env if it exists. Unreachable code in tests has also been removed, and .env is now in the .gitignore file.