Add coverage report to GitHub workflow (#3)
* Add coverage report to GitHub workflow

This commit introduces a new step to the GitHub actions workflow that generates a code coverage report. The change also includes adjustments to the Makefile to include coverage commands, and the .envrc file has been tweaked to interact with .env if it exists. Unreachable code in tests has also been removed, and .env is now in the .gitignore file.

* Add Codacy badges to README

Two Codacy badges have been added into the README file. These badges show the Codacy grade and the coverage of the code respectively, helping users to better understand the quality of the code.