Add GNU General Public License v3 (#8)
* Add GNU General Public License v3

A GNU General Public License Version 3 file has been added to the project. This provides the legal framework for the project, defining it as free software and outlining the terms of its distribution and modification.

* Add badges to README

This update includes the addition of several badges to the README.md file. These badges provide an at-a-glance look at the version information, Go reference, and Go version compatibility for the godash project.