Add Find function to godash package
The Find function has been added to the godash package and the test file 'example_slice_test.go'. The function will behave exactly like the [Find] function, except it is called directly on the slice to be searched. The related tests have also been written and included in the 'example_slice_test.go' file.