Creates Slice receiver methods for the possible functions  (#15)
* Add filter method and corresponding tests

This commit introduces a filter method in the godash.go file and updates the godash_test.go file to include corresponding tests. The filter method enhances functionality by allowing the use of a predicate function across elements in a given slice, returning a new slice with elements that pass the test.

* Add Filter function and update README

A new function, Filter, has been added to the examples_test.go and godash.go files to handle the filtering of slices. The README.md file has also been updated, including a new section providing a link to the documentation. The package description in godash.go has been slightly reformatted to improve clarity.

* Refactor test examples and update method documentation

Moved some test examples to a new `example_slice_test.go` file and removed them from the `examples_test.go` file. Updated documentation for the `Every` and `Filter` methods in `godash.go` to be more narrowly focused on their method behaviors, including usage examples.

* Add FindIndex method and related tests

Added a FindIndex method to the Slice struct in godash.go file to improve functionality. This method behaves exactly like the FindIndex function but is called directly on the slice to be searched. Corresponding tests have also been added in godash_test.go and example_slice_test.go to ensure its proper operation.

* Add Find function to godash package

The Find function has been added to the godash package and the test file 'example_slice_test.go'. The function will behave exactly like the [Find] function, except it is called directly on the slice to be searched. The related tests have also been written and included in the 'example_slice_test.go' file.

* Update minimum Go version to 1.19

The minimum Go version required for running tests and the project has been updated to 1.19. This is reflected in the tests.yaml workflow, README.md, and go.mod files. This change ensures compatibility with updated Go language features and improvements.