Creates a Set type (#18)
* Add basic Set implementation with tests

Introduced a generic Set type with methods for adding elements and retrieving all values. Added corresponding tests and an example to demonstrate functionality.

* Add String method to Set struct

Implemented a method to return a string representation of the set. Elements are sorted and formatted into a string, aiding in better readability and troubleshooting.

* Remove Go 1.19 and 1.20 from test matrix.

This change updates the unit testing workflow configuration to only include Go versions 1.21 and 1.22. It ensures the CI pipeline focuses on more recent Go versions, potentially reducing maintenance overhead.

* Update Go version to 1.22 in go.mod

This commit updates the Go version from 1.19 to 1.22 in the go.mod file. It ensures compatibility with newer Go features and improvements. The dependency on github.com/fatih/color remains unchanged.

* Refactor String method signature

Changed the receiver from a pointer receiver to a value receiver in the String method of the Set struct to ensure consistent behavior and potentially improve performance.

* Add Delete method to Set and corresponding tests

Implemented the Delete method to allow removal of elements from the Set. Added comprehensive unit tests to verify the behavior of deleting existing, non-existing, and elements from an empty set.

* Add Delete method example to set test

Introduce a new test illustrating the Set.Delete method. This enhances the test coverage and provides a practical example of using the Delete function.

* Add Has method and corresponding tests to Set

Implemented the Has method to check for the presence of an element in the Set. Added unit tests for the Has method to ensure it works correctly for various scenarios, including existing elements, non-existent elements, and an empty set.

* Add Has method to Set

Introduce the Has method to check for the existence of an element in the set. This method returns true if the element is found, and false otherwise.